Split vca/vcs into own files
